things starting to work, errors with DB delete

+++ src/start_mq_server.py
@@ -1,13 +1,14 @@
 from movie_quotes.framework.server.server import PieServer
 from movie_quotes.main import app
-from movie_quotes.framework.server import database
+from movie_quotes.framework import database
 import os
 
-os.remove("movie_quotes.db")
+if os.path.isfile("movie_quotes.db"):
+    os.remove("movie_quotes.db")
 
 conn , c = database.connect("movie_quotes.db")
-c.createTable("moviequotes", ["movie","quote"], c)
-c.createEntry("moviequotes", ["movie","quote"], ["tyler", "i dont care"], c)
+database.createTable("moviequotes", ["movie","quote"], c)
+database.createEntry("moviequotes", ["movie","quote"], ["tyler", "i dont care"], c)
 database.close(conn)
 
 PieServer(app).run()
